consolidated the duplicate pages folder in the react app
